Rename files to indicate the Redis type
