+++ examples/redis-lettuce/src/main/java/com/giffing/bucket4j/spring/boot/starter/RedisLettuceApplication.java
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class RedisApplication {
+public class RedisLettuceApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(RedisApplication.class, args);
+        SpringApplication.run(RedisLettuceApplication.class, args);
     }
 
 }
